% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{qpmadParameters}
\alias{qpmadParameters}
\title{Set qpmad parameters}
\usage{
qpmadParameters(
  isFactorized = FALSE,
  maxIter = -1,
  tol = 1e-12,
  checkPD = TRUE
)
}
\arguments{
\item{isFactorized}{If \code{TRUE} then \code{H} is a lower Cholesky factor.}

\item{maxIter}{Maximum number of iterations, if not positive then no limit.}

\item{tol}{Convergence tolerance.}

\item{checkPD}{If \code{FALSE} then \code{H} is assumed to be positive definite and no checks are made.}
}
\value{
a list suitable to be used as the pars-argument to \code{\link{solveqp}}
}
\description{
Set qpmad parameters
}
\examples{

qpmadParameters(checkPD = TRUE)

}
\seealso{
\code{\link{solveqp}}
}
